// src/types/shipper/transport.ts
export type ITransportMatching = {
  id: number;
  trans_type: number;
  created_user: string;
  created_date: string;
  updated_user: string;
  updated_date: string;
  carrier_operator_id: string;
  shipper_operator_id: string;
  carrier_operator_code: string;
  shipper_operator_code: string;
  shipper_operator_name: string;
  carrier_operator_name: string;
  carrier2_operator_id: string | null;
  carrier2_operator_code: string | null;
  batch_id: number;
  cns_line_item_by_date_id: number;
  vehicle_avb_resource_item_id: number;
  trailer_number: number;
  departure_from: number;
  arrival_to: number;
  transport_date: string;
  transport_code: string;
  status: number;
  request_snapshot: {
    id: number;
    status: number;
    weight: number | null;
    arrival_to: number;
    trans_type: number;
    operator_id: string;
    total_width: number | null;
    total_height: number | null;
    total_length: number | null;
    operator_code: string | null;
    departure_from: number;
    price_per_unit: number;
    trailer_number: number;
    trans_order_id: string | null;
    transport_code: string;
    transport_name: string;
    collection_date: string;
    cns_line_item_id: number | null;
    transport_plan_id: number | null;
    trans_plan_id: string | null;
    vehicle_condition: string | null;
    collection_time_to: string;
    outer_package_code: number;
    trailer_number_rest: number;
    collection_time_from: string;
    req_cns_line_item_id: number | null;
    special_instructions: string | null;
    temperature_range: string[];
  };
  propose_snapshot: {
    id: number;
    day: string;
    giai: string | null;
    price: number;
    status: number;
    trip_name: string;
    arrival_to: number;
    cut_off_time: number;
    cut_off_fee: number;
    cut_off_info_id: number;
    assign_type: string | null;
    max_payload: number | null;
    operator_id: string;
    total_width: number;
    arrival_time: string;
    arrival_date: string;
    total_height: number;
    total_length: number;
    vehicle_size: string | null;
    vehicle_type: string | null;
    display_order: number | null;
    operator_code: string | null;
    departure_from: number;
    departure_date: string;
    adjustment_price: number | null;
    temperature_range: string[];
    departure_time_max: string;
    departure_time_min: string;
    vehicle_diagram_id: string | null;
    vehicle_avb_resource_id: string | null;
    car_max_load_capacity1_meas: string | null;
    vehicle_name: string;
  };
  cnsg_prty_name_txt: string | null;
  cnsg_sct_sped_org_name_txt: string | null;
  cnsg_tel_cmm_cmp_num_txt: string | null;
  cnsg_pstl_adrs_line_one_txt: string;
  cnsg_pstc_cd: string | null;
  logs_srvc_prv_prty_name_txt: string;
  logs_srvc_prv_sct_sped_org_name_txt: string | null;
  logs_srvc_prv_sct_tel_cmm_cmp_num_txt: string | null;
  item_name_txt: string;
  trsp_cli_prty_name_txt: string;
  vehicle_avb_resource_item_id_created_at: string;
  trsp_cli_tel_cmm_cmp_num_txt: string;
  road_carr_depa_sped_org_name_txt: string;
  total: number;
  trailer: {
    service_no: string;
    service_name: string;
    service_strt_date: string;
    service_strt_time: string;
    service_end_date: string;
    service_end_time: string;
    freight_rate: number;
    car_ctrl_num_id: number | null;
    trailer_license_plt_num_id: string;
    departure_from: string;
    arrival_to: string;
  };
};